refactor: 영화 목록을 가져오는 함수 추가
